Updating documentation to explain duplicate key deserialization behaviour
This minor change is related to
https://github.com/aws/event-ruler/issues/22.
Would like to make sure there's at least something in the document
on the expected behaviour.

In future, this section can be updated with recommendation to
move to better APIs.

Also updating the pom.xml to make it easier to release changes. Going
forward, we'll enforce version bumps for key changes.