use fastdoubleparser to improve double parsing speed (#54)
Using FastDoubleParser improves convering to double. We do not see full benefits yet because of inherent issues within the subsequent code but this was a relatively small change to improve eventual performance. 

No new dependencies were added as we're reusing code that Jackson uses under the cover. Also the reason we feel comfortable with changing this API.