Making deletePatternRule a public method. (#80)
* Making deletePatternRule a public method.

We already have `addPatternRule` as public but offer no equivivalent
for deleting patterns. The method is tested as part of our
MachineTests.java so assuming we missed making this public in past
(probably because nobody asked until https://github.com/aws/event-ruler/issues/64).

Benchmarks were not necessary for this change.

* Supressing build failures becoase of docs. Will fix these later

* Update README to link  deteleRule() caveats

based on feedback on the pull-request.