Fixing CIDR matching to use floor of provided range (#89)
Previously, we would treat CIDR (of the form IP/bits) as a range from IP to maxBytes(IP, bits). But everything I read about CIDR notation, and from online calculators, says that a CIDR covers the range from minBytes(IP, bits) to maxBytes(IP, bits). I.e. The number of specified leading bits are fixed and the rest are variable. So 255.255.255.127/24 means the whole last 8 bits, or last byte, or "127", is variable, so the CIDR range should be 255.255.255.0 to 255.255.255.255, not 255.255.255.127 to 255.255.255.255.

* Fixing CIDR matching to use floor of provided range

* Fixing minor formatting issues

* Tweaking tests