Fix testPatternPermutations within ByteMachingTest (#182)
* Fix testPatternPermutations within ByteMachingTest

For each test within the class we add a list of "noMatches" objects
that should never be pattern matches intentinoally or not. However
we never check if we ever accidentally register these pattern when
adding other patterns. We will fix this by tracking when these
patterns get added (probably by error) and then later validate the
checks

Note, no impact to the correctness of the ByteMachine itself, just
that the test scaffolding was broken and needed to be fixed.

See https://github.com/aws/event-ruler/issues/181 for more context.

* Remover unused import