Optimization: removed level of indirection for SubRuleContext. (#173)
Instead of Double id use SubRuleContext directly everywhere. SubRuleContext id is only used internally for backward compatibility in equals and hashCode (not sure if this is actually needed, but tests check that). The benchmarks are not stable (need JMH with forking, proper warmup and multiple longer iterations for individual benchmarks), but overall can see 3-8% in throughput improvement on some benchmarks with this easy fix.

Using long as id instead of double because long is typically faster and actually can produce more unique 64-bit patterns than double because for long every 64-bit pattern is valid while for double it is not true (NaN).

Changed only types but did not rename variables all over the place because it would be harder to review, can be done later as needed.

https://github.com/aws/event-ruler/pull/173